+++ examples/pages/config.js
@@ -1,28 +1,8 @@
 export default [{
-    title: 'avue-cli 2.x 源码(html版)',
-    money: '99',
-    list: [{
-        title: 'html版本的cli2后台模版',
-        bold: true
-    }, {
-        title: '没有vue-cli那样子复杂的开发环境',
-        bold: true
-    }, {
-        title: '不依赖node开发环境，传统的html页面开发',
-        bold: true
-    }, {
-        title: '无任何协议限制,可二次开发',
-        bold: true
-    }, {
-        title: '私人交流学习群'
-    }, {
-        title: '前端最新干货分享'
-    }]
-}, {
     title: 'avue2.x源码',
-    money: '299',
+    money: '299.9',
     list: [{
-        title: 'avue核心源码解读<a href="https://www.npmjs.com/package/@smallwei/avue" target="_blank">https://cli2.avue.top</a>',
+        title: 'avue核心源码解读',
         bold: true
     }, {
         title: '大量的可配置组件源码解读',
@@ -44,15 +24,12 @@ export default [{
     }, {
         title: '前端最新干货分享'
     }, {
-        title: '免费的私人git私服<a href="https://git.avue.top" target="_blank">https://git.avue.top</a>'
+        title: '免费的私人git私服'
     }]
 }, {
     title: 'avue-cli 2.x源码',
-    money: '299',
+    money: '299.9',
     list: [{
-        title: '在线demo<a href="https://cli2.avue.top" target="_blank">https://cli2.avue.top</a>',
-        bold: true
-    }, {
         title: '面向全屏幕尺寸的响应式适配能力',
         bold: true
     }, {
@@ -74,7 +51,7 @@ export default [{
     }, {
         title: '模块的可拆卸化,达到开箱即用'
     }, {
-        title: '免费的私人git私服<a href="https://git.avue.top" target="_blank">https://git.avue.top</a>'
+        title: '免费的私人git私服'
     }, {
         title: '解放重复crud和form的枯燥劳动'
     }, {
@@ -83,9 +60,8 @@ export default [{
         title: '前端最新干货分享'
     }]
 }, {
-    span: 16,
     title: 'avue+avue-cli 2.x 源码',
-    money: '399',
+    money: '399.9',
     list: [{
         title: '同时拥有“avue2.x”和“avue-cli2.x”的授权',
         bold: true,
@@ -98,9 +74,11 @@ export default [{
     }, {
         title: '全家桶包含avue2.x和avue-cli2.x源码'
     }, {
-        title: '全家桶包含avue-form表单拖拽工具源码<a href="https://form.avue.top" target="_blank">https://form.avue.top</a>'
+        title: '全家桶包含avue-form表单拖拽工具源码'
     }, {
-        title: '全家桶包含avue-cli2html版本的的源码<a href="https://temp.avue.top" target="_blank">https://temp.avue.top</a>'
+        title: '全家桶包含avue-pig最新avue-cli对接pigx和pig前台'
+    }, {
+        title: '全家桶包含avue-cli2html版本的的源码'
     }, {
         title: '全家桶包含avue-api接口管理工具源码'
     }, {